+++ pkgs/intl_translation/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.19.0-dev
   * Always generate null safe code, remove `null-safe` flag.
   * Add example for `he` locale.
+  * MessageExtraction: add set of error codes to be ignored in analysis.
 
 ## 0.18.1
   * Update analyzer dependency to `5.2.0`.